bridge-web: refactor index page
