+++ src/components/Status/index.ts
@@ -2,3 +2,4 @@ export * from './Connected/Connected';
 export * from './Disconnected/Disconnected';
 export * from './NoAccounts/NoAccounts';
 export * from './NotInstalled/NoInstalled';
+export * from './WrongChain/WrongChain';
